+++ src/PublicRestApi/Appointments/Common.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Dbp\CampusonlineApi\PublicRestApi\Appointments;
+
+class Common
+{
+    public const API_PATH = 'co/co-res-core/appointment/api';
+}